backend/requirements.txt
# Core Django Framework
Django==5.2.4
djangorestframework==3.16.0

# Database
psycopg2-binary==2.9.10

# API and CORS
django-cors-headers==4.7.0
djangorestframework-camel-case==1.4.2

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# AI and Chatbot Dependencies (Pinned to local environment)
google-generativeai==0.3.2
#langchain==0.1.0
#langchain-google-genai==0.0.8
#langchain-community==0.0.13
#langchain-groq==0.3.7
#python-decouple==3.8
python-dotenv
cryptography

# Database
sqlalchemy==2.0.35

# Graph Database (if using Neo4j features)
neomodel==5.3.3
django-neomodel==0.2.0

# HTTP Requests (for testing and external API calls)
requests==2.31.0

# Development and Testing Dependencies
# (Optional - can be moved to separate requirements-dev.txt)
# These are used in test files but not required for production
#langchain-groq==0.3.7
spacy==3.7.4
pydantic==2.7.4